feat: update namespaces
