--- Devantler.SecretManager.slnx
@@ -1,14 +1,0 @@
-<Solution>
-  <Configurations>
-    <Platform Name="Any CPU" />
-    <Platform Name="x64" />
-    <Platform Name="x86" />
-  </Configurations>
-  <Folder Name="/src/">
-    <Project Path="src/Devantler.SecretManager.Core/Devantler.SecretManager.Core.csproj" />
-    <Project Path="src/Devantler.SecretManager.SOPS.LocalAge/Devantler.SecretManager.SOPS.LocalAge.csproj" />
-  </Folder>
-  <Folder Name="/tests/">
-    <Project Path="tests/Devantler.SecretManager.SOPS.LocalAge.Tests/Devantler.SecretManager.SOPS.LocalAge.Tests.csproj" />
-  </Folder>
-</Solution>
