+++ DevantlerTech.SecretManager.slnx
@@ -1,0 +1,14 @@
+<Solution>
+  <Configurations>
+    <Platform Name="Any CPU" />
+    <Platform Name="x64" />
+    <Platform Name="x86" />
+  </Configurations>
+  <Folder Name="/src/">
+    <Project Path="src/DevantlerTech.SecretManager.Core/DevantlerTech.SecretManager.Core.csproj" />
+    <Project Path="src/DevantlerTech.SecretManager.SOPS.LocalAge/DevantlerTech.SecretManager.SOPS.LocalAge.csproj" />
+  </Folder>
+  <Folder Name="/tests/">
+    <Project Path="tests/DevantlerTech.SecretManager.SOPS.LocalAge.Tests/DevantlerTech.SecretManager.SOPS.LocalAge.Tests.csproj" />
+  </Folder>
+</Solution>
