+++ src/DevantlerTech.SecretManager.Core/DevantlerTech.SecretManager.Core.csproj
@@ -12,14 +12,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Devantler.AgeCLI" Version="1.4.26" />
-    <PackageReference Include="Devantler.SOPSCLI" Version="1.6.18" />
-    <PackageReference Include="YAMLDotNet" Version="16.3.0" />
-    <PackageReference Include="Devantler.Commons.Utils" Version="0.6.4" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="../Devantler.SecretManager.Core/Devantler.SecretManager.Core.csproj" />
+    <PackageReference Include="DevantlerTech.Keys.Core" Version="1.0.30" />
   </ItemGroup>
 
 </Project>
