+++ src/DevantlerTech.SecretManager.SOPS.LocalAge/DevantlerTech.SecretManager.SOPS.LocalAge.csproj
@@ -12,7 +12,14 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Devantler.Keys.Core" Version="1.0.30" />
+    <PackageReference Include="DevantlerTech.AgeCLI" Version="1.4.26" />
+    <PackageReference Include="DevantlerTech.SOPSCLI" Version="1.6.18" />
+    <PackageReference Include="YAMLDotNet" Version="16.3.0" />
+    <PackageReference Include="DevantlerTech.Commons.Utils" Version="0.6.4" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="../DevantlerTech.SecretManager.Core/DevantlerTech.SecretManager.Core.csproj" />
   </ItemGroup>
 
 </Project>
