+++ tests/DevantlerTech.SecretManager.SOPS.LocalAge.Tests/DevantlerTech.SecretManager.SOPS.LocalAge.Tests.csproj
@@ -27,7 +27,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Devantler.SecretManager.SOPS.LocalAge\Devantler.SecretManager.SOPS.LocalAge.csproj" />
+    <ProjectReference Include="..\..\src\DevantlerTech.SecretManager.SOPS.LocalAge\DevantlerTech.SecretManager.SOPS.LocalAge.csproj" />
   </ItemGroup>
 
 </Project>
