+++ tests/DevantlerTech.SecretManager.SOPS.LocalAge.Tests/Utils/SOPSConfigHelperTests/CreateSOPSConfigAsyncTests.CreateSOPSConfigAsync_CreatesDirectoryAndConfigFile.verified.txt
