+++ tests/DevantlerTech.SecretManager.SOPS.LocalAge.Tests/Utils/SOPSConfigHelperTests/GetSOPSConfigAsyncTests.GetSOPSConfigAsync_GivenValidConfigPath_ReturnsSOPSConfig.verified.txt
